Add filament resource for tokens; Ignore non-personal tokens in API token endpoints

+++ app/Http/Controllers/Api/V1/ApiTokenController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Api\V1;
 
+use App\Exceptions\Api\PersonalAccessClientIsNotConfiguredException;
 use App\Http\Requests\V1\ApiToken\ApiTokenStoreRequest;
 use App\Http\Resources\V1\ApiToken\ApiTokenCollection;
 use App\Http\Resources\V1\ApiToken\ApiTokenWithAccessTokenResource;
@@ -26,7 +27,9 @@ class ApiTokenController extends Controller
     {
         $user = $this->user();
 
-        $tokens = $user->tokens()->get();
+        $tokens = $user->tokens()
+            ->where('client_id', '=', config('passport.personal_access_client.id'))
+            ->get();
 
         return new ApiTokenCollection($tokens);
     }
@@ -39,11 +42,15 @@ class ApiTokenController extends Controller
      *
      * @operationId createApiToken
      *
-     * @throws AuthorizationException
+     * @throws AuthorizationException|PersonalAccessClientIsNotConfiguredException
      */
     public function store(ApiTokenStoreRequest $request): ApiTokenWithAccessTokenResource
     {
         $user = $this->user();
+
+        if (config('passport.personal_access_client.id') === null || config('passport.personal_access_client.secret') === null) {
+            throw new PersonalAccessClientIsNotConfiguredException;
+        }
 
         $token = $user->createToken($request->getName(), ['*']);
         /** @var Token $tokenModel */
@@ -58,13 +65,20 @@ class ApiTokenController extends Controller
      * @operationId revokeApiToken
      *
      * @throws AuthorizationException
+     * @throws PersonalAccessClientIsNotConfiguredException
      */
     public function revoke(Token $apiToken): JsonResponse
     {
         $user = $this->user();
 
+        if (config('passport.personal_access_client.id') === null || config('passport.personal_access_client.secret') === null) {
+            throw new PersonalAccessClientIsNotConfiguredException;
+        }
         if ($apiToken->user_id !== $user->getKey()) {
             throw new AuthorizationException('API token does not belong to user');
+        }
+        if ($apiToken->client_id !== config('passport.personal_access_client.id')) {
+            throw new AuthorizationException('API token is not a personal access token');
         }
 
         $apiToken->revoke();
@@ -77,14 +91,20 @@ class ApiTokenController extends Controller
      *
      * @operationId deleteApiToken
      *
-     * @throws AuthorizationException
+     * @throws AuthorizationException|PersonalAccessClientIsNotConfiguredException
      */
     public function destroy(Token $apiToken): JsonResponse
     {
         $user = $this->user();
 
+        if (config('passport.personal_access_client.id') === null || config('passport.personal_access_client.secret') === null) {
+            throw new PersonalAccessClientIsNotConfiguredException;
+        }
         if ($apiToken->user_id !== $user->getKey()) {
             throw new AuthorizationException('API token does not belong to user');
+        }
+        if ($apiToken->client_id !== config('passport.personal_access_client.id')) {
+            throw new AuthorizationException('API token is not a personal access token');
         }
 
         $apiToken->delete();
